refactor(settings): extract shared Section/SettingItem primitives
把两个设置页共用的 Section / SettingItem 和 chip 样式抽到
components/SettingItem.tsx，appearance.tsx 改为 import，并清掉本地副本。
顺带把「本地 Feed 去重 / 启动时保留推荐流 / 清除本地去重记录」三项从
「实验性功能」迁出（迁入新增的「过滤与推荐」页），保留 WebView/私信/
浏览历史三项于原处。
